Reverting the moved types
