+++ src/eosjs-api-interfaces.ts
@@ -3,7 +3,7 @@
  * copyright defined in eosjs/LICENSE.txt
  */
 
-import { Abi, PushTransactionArgs } from './eosjs-rpc-interfaces';
+import { Abi, PushTransactionArgs, ProcessedAction } from './eosjs-rpc-interfaces';
 import { Anyvar, Authorization, Action, SerializedAction } from './eosjs-serialize';
 
 /** Arguments to `getRequiredKeys` */
@@ -100,6 +100,76 @@ export interface TransactionHeader {
     expiration: string;
     ref_block_num: number;
     ref_block_prefix: number;
+}
+
+export interface AccountDelta {
+    account: string;
+    delta: number;
+}
+
+export interface AuthSequence {
+    account: string;
+    sequence: number;
+}
+
+export interface ActionReceipt {
+    receiver: string;
+    act_digest: string;
+    global_sequence: number;
+    recv_sequence: number;
+    auth_sequence: [ string, number ][];
+    code_sequence: number;
+    abi_sequence: number;
+}
+
+export interface ActionTrace {
+    action_ordinal: number;
+    creator_action_ordinal: number;
+    closest_unnotified_ancestor_action_ordinal: number;
+    receipt: ActionReceipt;
+    receiver: string;
+    act: ProcessedAction;
+    context_free: boolean;
+    elapsed: number;
+    console: string;
+    trx_id: string;
+    block_num: number;
+    block_time: string;
+    producer_block_id: string|null;
+    account_ram_deltas: AccountDelta[];
+    account_disk_deltas: AccountDelta[];
+    except: any;
+    error_code: number|null;
+    return_value?: any;
+    return_value_hex_data?: string;
+    return_value_data?: any;
+    inline_traces?: ActionTrace[];
+}
+
+export interface TransactionReceiptHeader {
+    status: string;
+    cpu_usage_us: number;
+    net_usage_words: number;
+}
+
+export interface TransactionTrace {
+    id: string;
+    block_num: number;
+    block_time: string;
+    producer_block_id: string|null;
+    receipt: TransactionReceiptHeader|null;
+    elapsed: number;
+    net_usage: number;
+    scheduled: boolean;
+    action_traces: ActionTrace[];
+    account_ram_delta: AccountDelta|null;
+    except: string|null;
+    error_code: number|null;
+}
+
+export interface TransactResult {
+    transaction_id: string;
+    processed: TransactionTrace;
 }
 
 /** Optional query configuration object */
