+++ src/eosjs-rpc-interfaces.ts
@@ -4,6 +4,7 @@
  */
 
 import { Authorization } from './eosjs-serialize';
+import { TransactionReceiptHeader } from './eosjs-api-interfaces';
 
 /** Structured format for abis */
 export interface Abi {
@@ -479,76 +480,6 @@ export interface GetTableByScopeResultRow {
 export interface GetTableByScopeResult {
     rows: GetTableByScopeResultRow[];
     more: string;
-}
-
-export interface AccountDelta {
-    account: string;
-    delta: number;
-}
-
-export interface AuthSequence {
-    account: string;
-    sequence: number;
-}
-
-export interface ActionReceipt {
-    receiver: string;
-    act_digest: string;
-    global_sequence: number;
-    recv_sequence: number;
-    auth_sequence: [ string, number ][];
-    code_sequence: number;
-    abi_sequence: number;
-}
-
-export interface ActionTrace {
-    action_ordinal: number;
-    creator_action_ordinal: number;
-    closest_unnotified_ancestor_action_ordinal: number;
-    receipt: ActionReceipt;
-    receiver: string;
-    act: ProcessedAction;
-    context_free: boolean;
-    elapsed: number;
-    console: string;
-    trx_id: string;
-    block_num: number;
-    block_time: string;
-    producer_block_id: string|null;
-    account_ram_deltas: AccountDelta[];
-    account_disk_deltas: AccountDelta[];
-    except: any;
-    error_code: number|null;
-    return_value?: any;
-    return_value_hex_data?: string;
-    return_value_data?: any;
-    inline_traces?: ActionTrace[];
-}
-
-export interface TransactionReceiptHeader {
-    status: string;
-    cpu_usage_us: number;
-    net_usage_words: number;
-}
-
-export interface TransactionTrace {
-    id: string;
-    block_num: number;
-    block_time: string;
-    producer_block_id: string|null;
-    receipt: TransactionReceiptHeader|null;
-    elapsed: number;
-    net_usage: number;
-    scheduled: boolean;
-    action_traces: ActionTrace[];
-    account_ram_delta: AccountDelta|null;
-    except: string|null;
-    error_code: number|null;
-}
-
-export interface TransactResult {
-    transaction_id: string;
-    processed: TransactionTrace;
 }
 
 /** Arguments for `push_transaction` */
